+++ libs/galley-types/src/Galley/Types/Teams.hs
@@ -28,6 +28,7 @@ module Galley.Types.Teams
     flagLegalHold,
     flagTeamSearchVisibility,
     flagAppLockDefaults,
+    flagClassifiedDomains,
     Defaults (..),
     FeatureSSO (..),
     FeatureLegalHold (..),
@@ -194,7 +195,8 @@ data FeatureFlags = FeatureFlags
   { _flagSSO :: !FeatureSSO,
     _flagLegalHold :: !FeatureLegalHold,
     _flagTeamSearchVisibility :: !FeatureTeamSearchVisibility,
-    _flagAppLockDefaults :: !(Defaults (TeamFeatureStatus 'TeamFeatureAppLock))
+    _flagAppLockDefaults :: !(Defaults (TeamFeatureStatus 'TeamFeatureAppLock)),
+    _flagClassifiedDomains :: !(TeamFeatureStatus 'TeamFeatureClassifiedDomains)
   }
   deriving (Eq, Show, Generic)
 
@@ -237,14 +239,16 @@ instance FromJSON FeatureFlags where
       <*> obj .: "legalhold"
       <*> obj .: "teamSearchVisibility"
       <*> (fromMaybe (Defaults defaultAppLockStatus) <$> (obj .:? "appLock"))
+      <*> (fromMaybe defaultClassifiedDomains <$> (obj .:? "classifiedDomains"))
 
 instance ToJSON FeatureFlags where
-  toJSON (FeatureFlags sso legalhold searchVisibility appLock) =
+  toJSON (FeatureFlags sso legalhold searchVisibility appLock classifiedDomains) =
     object $
       [ "sso" .= sso,
         "legalhold" .= legalhold,
         "teamSearchVisibility" .= searchVisibility,
-        "appLock" .= appLock
+        "appLock" .= appLock,
+        "classifiedDomains" .= classifiedDomains
       ]
 
 instance FromJSON FeatureSSO where
@@ -284,6 +288,9 @@ makeLenses ''FeatureFlags
 -- The problem: the mapping between 'Role' and 'Permissions' is fixed by external contracts:
 -- client apps treat permission bit matrices as opaque role identifiers, so if we add new
 -- permission flags, things will break there.
+--
+-- "Hidden" in "HiddenPerm", therefore, refers to a permission hidden from
+-- clients, thereby making it internal to the backend.
 --
 -- The solution: add new permission bits to 'HiddenPerm', 'HiddenPermissions', and make
 -- 'hasPermission', 'mayGrantPermission' polymorphic.  Now you can check both for the hidden
@@ -329,8 +336,9 @@ roleHiddenPermissions role = HiddenPermissions p p
           [ ChangeLegalHoldTeamSettings,
             ChangeLegalHoldUserSettings,
             ChangeTeamSearchVisibility,
-            ChangeTeamFeature TeamFeatureAppLock {- the features not listed here can only be changed in stern -},
+            ChangeTeamFeature TeamFeatureAppLock,
             ChangeTeamFeature TeamFeatureFileSharing,
+            ChangeTeamFeature TeamFeatureClassifiedDomains {- the features not listed here can only be changed in stern -},
             ReadIdp,
             CreateUpdateDeleteIdp,
             CreateReadDeleteScimToken,
@@ -348,6 +356,7 @@ roleHiddenPermissions role = HiddenPermissions p p
           ViewTeamFeature TeamFeatureDigitalSignatures,
           ViewTeamFeature TeamFeatureAppLock,
           ViewTeamFeature TeamFeatureFileSharing,
+          ViewTeamFeature TeamFeatureClassifiedDomains,
           ViewLegalHoldUserSettings,
           ViewTeamSearchVisibility
         ]